Refactorisation du code avec amélioration de fonctions des classes Rooms, Users et Friendships.

+++ api/src/main/java/com/discoodle/api/service/RoomService.java
@@ -6,8 +6,10 @@ import com.discoodle.api.repository.UserRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.ThreadLocalRandom;
 
 @Service
 @AllArgsConstructor
@@ -25,34 +27,43 @@ public class RoomService {
         Room finalRoom = roomRepository.save(room);
         for (Long room_member : room_members) {
             if (userRepository.existsById(room_member)) {
-                roomRepository.addNewMember(finalRoom.getRoom_id(),room_member);
+                roomRepository.addNewMember(finalRoom.getRoom_id(), room_member);
             }
         }
         return finalRoom;
     }
 
     public Optional<Room> removeMember(String room_id, Long user_id) {
-        if(roomRepository.existsById(room_id)) {
-            Optional<Room> room = roomRepository.findRoomByUUID(room_id);
+        if (roomRepository.existsById(room_id) && userRepository.existsById(user_id)) {
+            Optional<Room> room = roomRepository.findById(room_id);
             roomRepository.removeMember(room_id, user_id);
+            if (room.get().getUsers().isEmpty()) {
+                roomRepository.deleteById(room_id);
+            }
+            if (room.get().getRoom_admin().equals(user_id)) {
+                room.get().setRoom_admin(room.get().getUsers().get(ThreadLocalRandom.current().nextInt(0, room.get().getUsers().size())).getId());
+            }
             return room;
         }
         return Optional.empty();
     }
 
     public void addNewMember(String room_id, Long user_id) {
-        roomRepository.addNewMember(room_id, user_id);
+        Optional<User> test = userRepository.findById(user_id);
+        if (test.isPresent() && !roomRepository.findById(room_id).get().getUsers().contains(test)) {
+            roomRepository.addNewMember(room_id, user_id);
+        }
     }
 
     public Optional<Room> changeLinkPicture(String room_id, String link_to_avatar) {
-        if(roomRepository.changeLinkPicture(room_id, link_to_avatar) == 1) {
+        if (roomRepository.changeLinkPicture(room_id, link_to_avatar) == 1) {
             return roomRepository.findById(room_id);
         }
         return Optional.empty();
     }
 
     public Optional<Room> changeAdmin(String room_id, Long room_admin) {
-        if(roomRepository.changeAdmin(room_id, room_admin) == 1) {
+        if (roomRepository.changeAdmin(room_id, room_admin) == 1) {
             return roomRepository.findById(room_id);
         }
         return Optional.empty();
